# dramatic speed improvements and code cleanup