@Neywiny 2025-09-07
Another few seconds gone.
```
$ hyperfine ./adsb_rx.py --output=pipe
Benchmark 1: ./adsb_rx.py
  Time (mean ± σ):     41.554 s ±  0.973 s    [User: 21.434 s, System: 23.290 s]
  Range (min … max):   40.347 s … 43.642 s    10 runs
 ```
This code is now running ~2/3 the amount of time as the before these commits. I'm still working through the exact mechanics of how this correlation is working, but I am pretty sure this is a safe change to make.