@Neywiny 2025-10-05
@peternmuller 

Good catch. Looks like something's different from our installs where that's an error. I re-ran my benchmark with just bool and the average went up a fraction of a second (within variance), but the bounds got tighter (not sure if within variance). So, to me, these are functionally the same just like the note from NumPy said they would be. Changing to int8 was a tad slower which matches my initial development findings.